+++ Changelog.md
@@ -19,6 +19,7 @@ Compiler Features:
  * SMTChecker: Support mod.
  * SMTChecker: Support ``contract`` type.
  * SMTChecker: Support ``this`` as address.
+ * SMTChecker: Support address members.
  * Optimizer: Add rule for shifts by constants larger than 255 for Constantinople.
  * Optimizer: Add rule to simplify certain ANDs and SHL combinations
  * Yul: Adds break and continue keywords to for-loop syntax.